Change calculation to polynomial formula

+++ VCNL4000Adapter/ProximityEventArgs.cs
@@ -4,7 +4,7 @@ namespace VCNL4000Adapter
 {
     public class ProximtyEventArgs : EventArgs
     {
-        public decimal Proximity { get; set; }
+        public double Proximity { get; set; }
         public int RawValue { get; set; }
     }
 }
